// src/common/enum/entity/user/user-language.enum.ts
/**
 * Enumeration of operation types for authentication user actions.
 * Each member of this enum represents a specific operation type.
 */
export enum UserEntityEnum {
  LANGUAGE_SPANISH = 'es', // Represents the Spanish language.
  LANGUAGE_ENGLISH = 'en', // Represents the English language.
  LANGUAGE_PORTUGUESE = 'pt', // Represents the Portuguese language.
  LANGUAGE_FRENCH = 'fr', // Represents the French language.
  LANGUAGE_ITALIAN = 'it', // Represents the Italian language.
  LANGUAGE_GERMAN = 'de', // Represents the German language.
  LANGUAGE_CHINESE = 'zh', // Represents the Chinese language.
}